implement generic mocha tester
